class09_20251019/05_structured_output/pyproject.toml
[project]
name = "15-structured-output"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai==1.98.0",
    "openai-agents>=0.2.5",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
]
